requirements.txt
apache-flink==1.19.1
kafka-python==2.0.2